pyproject.toml
[tool.poetry]
name = "backend"
version = "1.0.0"
description = "backend"
authors = ["developer <developer@mail.com>"]


[tool.black]
line-length = 79


[tool.isort]
multi_line_output=3
include_trailing_comma=true
use_parentheses=true
line_length=79


[tool.taskipy.tasks]
manage = "python ./backend/manage.py"
clear = "task manage clear"
dumpdata = "task manage dump"
migrate = "task manage migrate --noinput"
server = "task manage runserver"
loaddata = "task manage loaddata"
startapp = "task manage createapp"
initconfig = "task manage initconfig"
makemigrations = "task manage makemigrations"
createsuperuser = "task manage createsuperuser"
collectstatic = "task manage collectstatic --noinput"
gunicorn = "gunicorn --timeout 300 --chdir ./backend config.wsgi:application -b 0.0.0.0:8000"
defaultadmin = "task manage defaultadmin"
defaultfixtures = "task manage defaultfixtures"
graph = "task manage graph_models --pydot -a -g -o graph/visualized.png"


[tool.poetry.dependencies]
python = "^3.7, <3.9"
django = "^3.0.3"
django-cleanup = "^4.0.0"
django-widget-tweaks = "^1.4.5"
django-settings-export = "^1.2.1"
django-webpack-loader = "^0.7.0"
djangorestframework = "^3.11.0"
django-cors-headers = "^3.2.1"
drf-yasg = "^1.17.0"
arrow = "^0.15.5"
django-import-export = "^2.0.1"
django-environ = "^0.4.5"
gunicorn = "^20.0.4"
mysqlclient = "^1.4.6"
psycopg2 = "^2.8.4"
pillow = "^7.0.0"
django-storages = "^1.9.1"
django-etc = "^1.2.0"
django-rest-multiple-models = "^2.1.3"
djangorestframework-camel-case = "^1.1.2"
django-extensions = "^3.0.2"
pyparsing = "^2.4.7"
pydot = "^1.4.1"
djangorestframework-recursive = "^0.1.2"
drf-writable-nested = "^0.6.0"
django-weasyprint = "^1.0.1"
tqdm = "^4.48.2"
xlsxwriter = "^1.3.3"
transliterate = "^1.10.2"
py-linq = "^1.2.4"
django-service-objects = "^0.7.0"
djangorestframework-simplejwt = {git = "https://github.com/igor2104/django-rest-framework-simplejwt.git", rev = "jwt_cookie"}
loguru = "^0.5.3"
django-ckeditor = "^6.1.0"
django-admin-interface = "^0.18.2"
django-apptemplates = "^1.5"
django-extra-settings = "^0.1.3"
django-easy-select2 = "^1.5.7"
django-modeltranslation = "^0.17.3"
captcha = "^0.3"
django-filter = "^21.1"




[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
django-debug-toolbar = "^2.2"
autopep8 = "^1.5"
pylint-django = "^2.0.13"
taskipy = "^1.1.3"
flake8 = "^3.7.9"
black = "^19.10b0"
yolk3k = "^0.9"

